# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API PetCare",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "https://petcare-production-cc6b.up.railway.app/api"
        },
        {
            "url": "http://127.0.0.1:8000/api"
        },
        {
            "url": "https://petcare-production-3202.up.railway.app/api"
        }
    ],
    "paths": {
...
        "/animals": {
            "get": {
                "tags": [
                    "User - Animals"
                ],
                "description": "Get all animals",
                "operationId": "get_ll_animals",
                "parameters": [
                    {
                        "name": "category_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "breed_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "gender",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "male",
                                "female"
                            ]
                        }
                    },
                    {
                        "name": "size",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "small",
                                "medium",
                                "large"
                            ]
                        }
                    },
                    {
                        "name": "health_status",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "healthy",
                                "injured",
                                "sick"
                            ]
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "with_paginate",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "enum": [
                                0,
                                1
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            },
            "post": {
                "tags": [
                    "User - Animals"
                ],
                "description": "Create new animal.",
                "operationId": "create_animal",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "name",
                                    "category_id",
                                    "breed_id",
                                    "primary_color",
                                    "address",
                                    "gender",
                                    "size",
                                    "profile_image",
                                    "cover_image",
                                    "health_status",
                                    "birth_date"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "category_id": {
                                        "type": "integer"
                                    },
                                    "breed_id": {
                                        "type": "integer"
                                    },
                                    "primary_color": {
                                        "type": "string"
                                    },
                                    "secondary_color": {
                                        "type": "string"
                                    },
                                    "address": {
                                        "type": "string"
                                    },
                                    "weight": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "gender": {
                                        "type": "string",
                                        "enum": [
                                            "male",
                                            "female"
                                        ]
                                    },
                                    "size": {
                                        "type": "string",
                                        "enum": [
                                            "small",
                                            "medium",
                                            "large"
                                        ]
                                    },
                                    "desc": {
                                        "type": "string"
                                    },
                                    "profile_image": {
                                        "type": "file"
                                    },
                                    "cover_image": {
                                        "type": "file"
                                    },
                                    "birth_date": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "health_status": {
                                        "type": "string",
                                        "enum": [
                                            "healthy",
                                            "injured",
                                            "sick",
                                            "unknown"
                                        ]
                                    },
                                    "attachments[0][title]": {
                                        "type": "string"
                                    },
                                    "attachments[0][file]": {
                                        "type": "file"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/animals/{id}": {
            "get": {
                "tags": [
                    "User - Animals"
                ],
                "description": "Get specific animal",
                "operationId": "show_animal",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "bearer_token": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}